backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
redis==5.2.0
falkordb==1.0.8

# AI/LLM
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
openai==1.54.3
tiktoken==0.8.0

# WebSocket
python-socketio==5.11.3
python-multipart==0.0.12

# Utils
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4